Revert "sab redux style optimization"
This reverts commit bdefc3a635337413eb9d09f4e86bd79546ec0ccb.

theres some good stuff here to takeaway on, impl messed up the output

+++ demo/src/level/types.hpp
@@ -10,6 +10,6 @@ namespace demo::level {
         u16       nColumns;     // level width in metatile columns
         const u8* DynData;      // dynamic plane: RLE run-data (ROM, copied to RAM at load)
         const u8* DynLengths;   // dynamic plane: RLE run-lengths (ROM)
-        u16       DynRuns;      // dynamic-plane run count to copy into the RAM pool
+        u8        DynRuns;      // dynamic-plane run count to copy into the RAM pool
     };
 }
